+++ src/sql/ZIMSqlSelectStatement.m
@@ -408,11 +408,11 @@
 	}
 	
 	if (_limit > 0) {
-		[sql appendFormat: @" LIMIT %u", _limit];
+		[sql appendFormat: @" LIMIT %lu", (unsigned long)_limit];
 	}
 
 	if (_offset > 0) {
-		[sql appendFormat: @" OFFSET %u", _offset];
+		[sql appendFormat: @" OFFSET %lu", (unsigned long)_offset];
 	}
 
 	for (NSString *combine in _combine) {
